% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_fix_Dep.R
\name{fix_Dep}
\alias{fix_Dep}
\title{Detect linear dependencies of one matrix on another}
\usage{
fix_Dep(X1, X2, tol = .Machine$double.eps^0.5, rank.def = 0, strict = FALSE)
}
\arguments{
\item{X1}{A matrix.}

\item{X2}{A matrix, the columns of which may be partially linearly dependent on the columns of X1.}

\item{tol}{The tolerance to use when assessing linear dependence.}

\item{rank.def}{If the degree of rank deficiency in X2, given X1, is known, then it can be supplied here, and tol is then ignored. Unused unless positive and not greater than the number of columns in X2.}

\item{strict}{if TRUE then only columns individually dependent on X1 are detected, if FALSE then enough columns to make the reduced X2 full rank and independent of X1 are detected.}

\item{The}{tolerance to use when assessing linear dependence.}
}
\value{
A vector of the columns of X2 which are linearly dependent on columns of X1 (or which need to be deleted to acheive independence and full rank if strict==FALSE). NULL if the two matrices are independent.
}
\description{
Detect linear dependencies of one matrix on another
}
