% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_apply.R
\name{GWASapply}
\alias{GWASapply}
\title{GWAS with PCA}
\usage{
GWASapply(
  pheno = NULL,
  geno = NULL,
  Cov = NULL,
  GM = NULL,
  PCA.M = 3,
  QTN.position = NULL,
  cutoff = NULL,
  plots = FALSE,
  messages = FALSE,
  print = FALSE,
  trait = "unknown"
)
}
\arguments{
\item{pheno}{file with numeric phenotypic values}

\item{geno}{data.frame with genotype calls coded as 0,1,2.}

\item{Cov}{numeric data.frame with covariates values}

\item{GM}{genetic map of data with chr and position of each SNP}

\item{PCA.M}{number of principal components to use default is 3}

\item{QTN.position}{posistion of QTN if applicable}

\item{cutoff}{If cutoff is default, uses Bonferroni; else uses -log(value) of 0.05/number of SNPs}

\item{plots}{if TRUE, function plots PCA graphs, Manhatten Plot and QQ plot}

\item{messages}{if TRUE, returns messages for the GWAS function}

\item{print}{if TRUE, results are saved in a CSV}

\item{trait}{character value for trait name}
}
\value{
Manhatten plot, QQ plot plus p-values, type-1 error and power for every SNP and results in a CSV
}
\description{
GWAS with PCA
}
