% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_test.R
\name{GWAStest}
\alias{GWAStest}
\title{GWAS with PCA}
\usage{
GWAStest(
  phenotypes = NULL,
  genotypes = NULL,
  Cov = NULL,
  GM = NULL,
  PCA.M = 3,
  QTN.position = NULL,
  cutoff = NULL
)
}
\arguments{
\item{phenotypes}{file with numeric phenotypic values}

\item{genotypes}{data.frame with genotype calls coded as 0,1,2.}

\item{Cov}{numeric data.frame with covariates values}

\item{GM}{genetic map of data with chr and position of each SNP}

\item{PCA.M}{number of principal components to use default is 3}

\item{QTN.position}{posistion of QTN if applicable}

\item{cutoff}{If cutoff is default, uses Bonferroni; else uses -log(value) of 0.05/number of SNPs}
}
\value{
Manhatten plot, QQ plot plus p-values, type-1 error and power for every SNP
}
\description{
GWAS with PCA
}
