% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_manhattan.R
\name{manhattan_plot}
\alias{manhattan_plot}
\title{Function to create a manhattan plot}
\usage{
manhattan_plot(
  GM,
  pvals,
  cutoff = NULL,
  QTN_index = c(),
  FP = NULL,
  TP = NULL,
  trait = "unknown",
  messages = FALSE
)
}
\arguments{
\item{GM}{genetic map of data with chr and position of each SNP}

\item{pvals}{pvals from gwas results for each SNP}

\item{cutoff}{If cutoff is default, uses Bonferroni; else uses -log(value) of 0.05/number of SNPs}

\item{QTN_index}{posistion of QTN if applicable}

\item{FP}{Positions of SNPS that are False positive}

\item{TP}{Positions of SNPS that are True positive}

\item{trait}{character value for trait name}
}
\value{
Manhatten plot
}
\description{
Function to create a manhattan plot
}
