% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helloroxygenize.R
\name{math}
\alias{math}
\title{Add together two numbers}
\usage{
math(x, y)
}
\arguments{
\item{x}{number}

\item{y}{another number}
}
\value{
The sum of both numbers
}
\description{
Add together two numbers
}
