% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helloroxygenize.R
\name{math2}
\alias{math2}
\title{Add together two numbers}
\usage{
math2(x, y, y1)
}
\arguments{
\item{x}{number}

\item{y}{another number}

\item{y1}{a third number}
}
\value{
The sum of all there numbers
}
\description{
Add together two numbers
}
