% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_Significance_Error_Tests.R
\name{power.fdr}
\alias{power.fdr}
\title{Function to caculate power, FDR and type-1 error, False Positives, and True Positives}
\usage{
power.fdr(P.value, QTN.position = NULL, cutoff = NULL)
}
\arguments{
\item{QTN.position}{position of QTN if known}

\item{P}{list of SNPs order by ascending  p-value}
}
\value{
list of power, FDR, type-1 error, False Positives, and True Positives
}
\description{
Function to caculate power, FDR and type-1 error, False Positives, and True Positives
}
