% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_powerFDRtype1error.R
\name{power.fdr}
\alias{power.fdr}
\title{Function to caculate power, FDR and type-1 error}
\usage{
power.fdr(order.SNP, QTN.position = NULL)
}
\arguments{
\item{order.SNP}{list of SNPs order by ascending  p-value}

\item{QTN.position}{position of QTN if known}
}
\value{
list of power, FDR and type-1 error
}
\description{
Function to caculate power, FDR and type-1 error
}
